+++ packages/compost/.gitignore
@@ -11,6 +11,7 @@ images
 *.gif
 *.svg
 *.webp
+*.heic
 *.ico
 *.txt
 *.pdf
