+++ packages/compost/entities/__init__.py
@@ -1,0 +1,12 @@
+"""Game entities: chunks and worms."""
+
+from entities.chunk import Chunk, GluedChunk
+from entities.worm import WormHistory, Glue, Worm
+
+__all__ = [
+    'Chunk',
+    'GluedChunk',
+    'WormHistory',
+    'Glue',
+    'Worm',
+]
