+++ packages/compost/managers/__init__.py
@@ -1,0 +1,15 @@
+"""Manager modules for simulation subsystems."""
+
+from managers.audio_manager import AudioManager
+from managers.boundary_manager import BoundaryManager
+from managers.export_manager import ExportManager
+from managers.queue_manager import QueueManager
+from managers.worm_manager import WormManager
+
+__all__ = [
+    'AudioManager',
+    'BoundaryManager',
+    'ExportManager',
+    'QueueManager',
+    'WormManager',
+]
