+++ packages/compost/network/__init__.py
@@ -1,0 +1,7 @@
+"""Network components for the simulation."""
+
+from network.upload_server import UploadServer
+
+__all__ = [
+    'UploadServer',
+]
