+++ packages/compost/ui/__init__.py
@@ -1,0 +1,7 @@
+"""UI components for the simulation."""
+
+from ui.ui_manager import UIManager
+
+__all__ = [
+    'UIManager',
+]
